fix all within page.client.tsx
